# example/pb/app.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "example/pb/app.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/keyval/{key}": {
      "get": {
        "operationId": "KeyValRead",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbKeyValMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "value",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "KeyVal"
        ]
      },
      "delete": {
        "operationId": "KeyValDelete",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbKeyValMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "KeyVal"
        ]
      },
      "post": {
        "operationId": "KeyValUpdate",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbKeyValMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbKeyValMessage"
            }
          }
        ],
        "tags": [
          "KeyVal"
        ]
      },
      "put": {
        "operationId": "KeyValCreate",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/pbKeyValMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbKeyValMessage"
            }
          }
        ],
        "tags": [
          "KeyVal"
        ]
      }
    }
  },
  "definitions": {
    "pbKeyValMessage": {
      "type": "object",
      "properties": {
        "key": {
          "type": "string"
        },
        "value": {
          "type": "string"
        }
      }
    }
  }
}